feat: add shared utility module
